call getUrl on the cacher rather than a new manager method
