+++ tests/StaticCaching/ApplicationCacherTest.php
@@ -177,4 +177,70 @@ class ApplicationCacherTest extends TestCase
         $this->assertEquals([], $cacher->getUrls('http://example.com')->all());
         $this->assertEquals([], $cacher->getUrls('http://another.com')->all());
     }
+
+    #[Test]
+    public function it_gets_the_current_url()
+    {
+        $request = Request::create('http://example.com/test', 'GET');
+
+        $this->assertEquals('http://example.com/test', (new ApplicationCacher(app(Repository::class), []))->getUrl($request));
+    }
+
+    #[Test]
+    public function it_gets_the_current_url_with_query_parameters()
+    {
+        $request = Request::create('http://example.com/test', 'GET', [
+            'foo' => 'bar',
+        ]);
+
+        $this->assertEquals('http://example.com/test?foo=bar', (new ApplicationCacher(app(Repository::class), []))->getUrl($request));
+    }
+
+    #[Test]
+    public function it_gets_the_current_url_with_query_strings_disabled()
+    {
+        config()->set('statamic.static_caching.ignore_query_strings', true);
+
+        $request = Request::create('http://example.com/test', 'GET', [
+            'foo' => 'bar',
+        ]);
+
+        $this->assertEquals('http://example.com/test', (new ApplicationCacher(app(Repository::class), []))->getUrl($request));
+    }
+
+    #[Test]
+    public function it_gets_the_current_url_with_allowed_query_parameters()
+    {
+        config()->set('statamic.static_caching.allowed_query_strings', [
+            'foo', 'quux',
+        ]);
+
+        $request = Request::create('http://example.com/test', 'GET', [
+            'foo' => 'bar',
+            'baz' => 'qux',
+            'quux' => 'corge',
+        ]);
+
+        $this->assertEquals('http://example.com/test?foo=bar&quux=corge', (new ApplicationCacher(app(Repository::class), []))->getUrl($request));
+    }
+
+    #[Test]
+    #[DataProvider('disallowedQueryParametersProvider')]
+    public function it_gets_the_current_url_with_disallowed_query_parameters(array $disallowed, string $url, array $query, string $expected)
+    {
+        config()->set('statamic.static_caching.disallowed_query_strings', $disallowed);
+
+        $request = Request::create($url, 'GET', $query);
+
+        $this->assertEquals($expected, (new ApplicationCacher(app(Repository::class), []))->getUrl($request));
+    }
+
+    public static function disallowedQueryParametersProvider()
+    {
+        return [
+            [[], 'http://example.com/test', ['foo' => 'bar'], 'http://example.com/test?foo=bar'],
+            [['quux'], 'http://example.com/test', ['quux' => 'corge'], 'http://example.com/test'],
+            [['quux'], 'http://example.com/test', ['foo' => 'bar', 'baz' => 'qux', 'quux' => 'corge'], 'http://example.com/test?baz=qux&foo=bar'],
+        ];
+    }
 }
